Move widgets into submodule of UI
